thcon-listen: optionally exit if pre-existing socket found
Some applications launch a new instance of an extension for each new
window, but have a settings-modification API that modifies all windows
(e.g. Atom [1]).  Since those windows don't have a concept of a
"primary" instance, it's easiest to launch `thcon-listen` and have it
exit immediately when a socket already exists for that application.  Add
`--no-replace` CLI argument to exit if a socket already exists.

[1] https://atom.io