config: allow apps to be disabled via config
Before any apps can have a default configuration, it needs to be
possible for users to opt out of switching themes in an app when
`thcon dark` or `thcon light` is run.  Allow `disabled = true`
in place of a fully-specified app configuration to skip an app when not
explicitly listed in argv.